Add Cython implementation to tests
